Fix Student and Tutor class fields and methods
